refactor: Refact releaseNote API match to convention
